tools/dvmcfggen/requirements.txt
pyyaml>=6.0
textual>=0.47.0
rich>=13.0.0
